14th as

+++ Project KYM/Assets/01_Project KYM/Scripts/Interface/IHittable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHittable 
 {
-    void OnHit(int damage);
+    void OnHit(float damage);
 }
