+++ Project KYM/Assets/01_Project KYM/Scripts/Misc/Switch.cs
@@ -19,7 +19,7 @@ namespace KYM
             }
         }
 
-        public void OnHit(int damage)
+        public void OnHit(float damage)
         {
             if (TargetObj != null)
             {
